# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UEX Merchant Platform API",
        "description": "REST API for UEX merchant cabinet",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current environment"
        }
    ],
    "paths": {
        "/api/merchant/auth/exchange": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Exchange backend JWT for merchant JWT",
                "description": "Class ExchangeTokenController",
                "operationId": "e6730ff3ad806cd65a1ddfce8c37f29f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Backend JWT issued by uexapp-backend",
                                        "type": "string",
                                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Merchant JWT issued successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 3600
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
        "/api/merchant/oauth2/token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Issue merchant JWT via client credentials (server-to-server)",
                "description": "Class OAuth2TokenController",
                "operationId": "35ea04863c233732e16397719da1584b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "client_secret",
                                    "grant_type"
                                ],
                                "properties": {
                                    "client_id": {
                                        "description": "Merchant app client_id from merchant_apps",
                                        "type": "string",
                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                    },
                                    "client_secret": {
                                        "description": "Merchant app client_secret",
                                        "type": "string",
                                        "example": "xK9mP2qR7tN4vL1wY8uZ3sA6bC0dE5fG"
                                    },
                                    "grant_type": {
                                        "description": "OAuth2 grant type — must be client_credentials",
                                        "type": "string",
                                        "example": "client_credentials"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Merchant JWT issued successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 3600
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "Unauthorized": {
                "description": "Unauthenticated or invalid token",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Validation failed",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Validation failed."
                                },
                                "errors": {
                                    "type": "object",
                                    "example": {
                                        "field": [
                                            "The field is required."
                                        ]
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}